Updated test environment * Not printing stderr to out file   * To prevent make warning when -j flag is used * Make out directory, kernel, and fs.img as dependency on make test   * To prevent make messages be written to out file * Kill qemu process when ^c is pressed on run-test.py * Added out directory to .gitignore * Added result comparing utility

+++ run-test.py
@@ -24,37 +24,45 @@ from subprocess import *
 
 def test_i(i):
   print("Test %d ... " % i, end="", flush=True)
-  qemu = Popen(["make", "-s", "qemu-verbose"], 
+  qemu = Popen(["make", "-s", "qemu-verbose"],
       stdin=PIPE, stdout=PIPE, stderr=PIPE,
       universal_newlines=True, preexec_fn=os.setsid)
+  qemu_killed = False
 
-  time.sleep(1)
-  qemu.stdin.write("kthtest %d\n" % i)
-  qemu.stdin.flush()
-  time.sleep(5)
-  qemu.stdin.write("\x01x");
-  qemu.stdin.flush()
-
-  results = tuple()
   try:
-    results = qemu.communicate(timeout=10)
-  except TimeoutExpired:
-    os.killpg(os.getpgid(qemu.pid), signal.SIGTERM)
-    results = qemu.communicate()
+    time.sleep(1)
+    qemu.stdin.write("kthtest %d\n" % i)
+    qemu.stdin.flush()
+    time.sleep(5)
+    qemu.stdin.write("\x01x");
+    qemu.stdin.flush()
 
-  out_path = Path("./out/kthtest%d.out" % i)
-  with out_path.open("wt") as outfd:
-    print(results[0], file=outfd)
-    print(results[1], file=outfd)
+    results = tuple()
+    try:
+      results = qemu.communicate(timeout=10)
+    except TimeoutExpired:
+      os.killpg(os.getpgid(qemu.pid), signal.SIGTERM)
+      results = qemu.communicate()
+    qemu_killed = True
 
-  
-for i in range(10):
-  test_i(i)
-  and_file = "./ans/kthtest%d.ans" % i
-  out_file = "./out/kthtest%d.out" % i
-  rc = os.system("cmp -s " + and_file + " " + out_file)
-  if rc == 0:
-    print("PASS")
-  else:
-    print("FAIL")
+    out_path = Path("./out/kthtest%d.out" % i)
+    with out_path.open("wt") as outfd:
+      print(results[0], file=outfd)
 
+  finally:
+    if not qemu_killed:
+      os.killpg(os.getpgid(qemu.pid), signal.SIGTERM)
+
+
+try:
+  for i in range(10):
+    test_i(i)
+    and_file = "./ans/kthtest%d.ans" % i
+    out_file = "./out/kthtest%d.out" % i
+    rc = os.system("cmp -s " + and_file + " " + out_file)
+    if rc == 0:
+      print("PASS")
+    else:
+      print("FAIL")
+except KeyboardInterrupt:
+  pass
